+++ src/text_face.rs
@@ -206,6 +206,7 @@ pub(crate) struct SpecifiedTextFace {
     pub(crate) bg: Option<Color>,
     pub(crate) underline_color: Option<Color>,
     pub(crate) style: Option<TextStyling>,
+    pub(crate) reset: bool,
 }
 
 pub(crate) fn parse_text_face(arguments: &[WString]) -> SpecifiedTextFace {
@@ -217,7 +218,7 @@ pub(crate) fn parse_text_face(arguments: &[WString]) -> SpecifiedTextFace {
     match parse_text_face_and_options(&mut argv, /*is_builtin=*/ false) {
         Ok(SetFace(specified_face)) => specified_face,
         Err(_) => Default::default(),
-        Ok(ResetFace) | Ok(PrintColors(_)) | Ok(PrintHelp) => unreachable!(),
+        Ok(PrintColors(_)) | Ok(PrintHelp) => unreachable!(),
     }
 }
 
@@ -230,7 +231,6 @@ pub(crate) struct PrintColorsArgs<'argarray, 'args> {
 
 pub(crate) enum ParsedArgs<'argarray, 'args> {
     SetFace(SpecifiedTextFace),
-    ResetFace,
     PrintHelp,
     PrintColors(PrintColorsArgs<'argarray, 'args>),
 }
@@ -242,6 +242,8 @@ pub(crate) enum ParseError<'args> {
     UnknownColor(&'args wstr),
     UnknownUnderlineStyle(&'args wstr),
     UnknownOption(usize),
+    InvalidFgArgCombination,
+    InvalidFgPrintColorCombination,
 }
 
 fn parse_resettable_style<'a>(w: &WGetopter<'_, 'a, '_>) -> Result<ResettableStyle, &'a wstr> {
@@ -260,9 +262,10 @@ pub(crate) fn parse_text_face_and_options<'argarray, 'args>(
     is_builtin: bool,
 ) -> Result<ParsedArgs<'argarray, 'args>, ParseError<'args>> {
     let builtin_extra_args = if is_builtin { 0 } else { "hc".len() };
-    let short_options = L!("b:oi::dr::s::u::ch");
+    let short_options = L!("f:b:oi::dr::s::u::ch");
     let short_options = &short_options[..short_options.len() - builtin_extra_args];
     let long_options: &[WOption] = &[
+        wopt(L!("foreground"), ArgType::RequiredArgument, 'f'),
         wopt(L!("background"), ArgType::RequiredArgument, 'b'),
         wopt(L!("underline-color"), ArgType::RequiredArgument, '\x02'),
         wopt(L!("bold"), ArgType::NoArgument, 'o'),
@@ -274,6 +277,7 @@ pub(crate) fn parse_text_face_and_options<'argarray, 'args>(
         wopt(L!("theme"), ArgType::RequiredArgument, '\x01'),
         wopt(L!("help"), ArgType::NoArgument, 'h'),
         wopt(L!("print-colors"), ArgType::NoArgument, 'c'),
+        wopt(L!("reset"), ArgType::NoArgument, '\x03'),
     ];
     let long_options = &long_options[..long_options.len() - builtin_extra_args];
 
@@ -291,6 +295,7 @@ pub(crate) fn parse_text_face_and_options<'argarray, 'args>(
         }
     };
 
+    let mut fg_colors = vec![];
     let mut bg_colors = vec![];
     let mut underline_colors = vec![];
     let mut style: Option<TextStyling> = None;
@@ -298,10 +303,16 @@ pub(crate) fn parse_text_face_and_options<'argarray, 'args>(
         style.get_or_insert(TextStyling::unknown_style())
     }
     let mut print_color_mode = false;
+    let mut reset = false;
 
     let mut w = WGetopter::new(short_options, long_options, argv);
     while let Some(c) = w.next_opt() {
         match c {
+            'f' => {
+                if let Some(fg) = parse_color(w.woptarg.unwrap())? {
+                    fg_colors.push(fg);
+                }
+            }
             'b' => {
                 if let Some(bg) = parse_color(w.woptarg.unwrap())? {
                     bg_colors.push(bg);
@@ -312,6 +323,9 @@ pub(crate) fn parse_text_face_and_options<'argarray, 'args>(
                 if let Some(underline_color) = parse_color(w.woptarg.unwrap())? {
                     underline_colors.push(underline_color);
                 }
+            }
+            '\x03' => {
+                reset = true;
             }
             'h' => {
                 assert!(is_builtin);
@@ -367,6 +381,9 @@ pub(crate) fn parse_text_face_and_options<'argarray, 'args>(
     }
 
     let fg_args = &w.argv[w.wopt_index..];
+    if !fg_args.is_empty() && !fg_colors.is_empty() {
+        return Err(InvalidFgArgCombination);
+    }
 
     let best_color =
         |colors: Vec<Color>| terminal::best_color(colors.into_iter(), get_color_support());
@@ -375,6 +392,9 @@ pub(crate) fn parse_text_face_and_options<'argarray, 'args>(
     let underline_color = best_color(underline_colors);
 
     if print_color_mode {
+        if !fg_colors.is_empty() {
+            return Err(InvalidFgPrintColorCombination);
+        }
         return Ok(PrintColors(PrintColorsArgs {
             fg_args,
             bg,
@@ -385,15 +405,21 @@ pub(crate) fn parse_text_face_and_options<'argarray, 'args>(
 
     // Historical behavior: reset only applies if it's the first argument.
     if is_builtin && fg_args.first().is_some_and(|fg| fg == "reset") {
-        return Ok(ResetFace);
+        return Ok(SetFace(SpecifiedTextFace {
+            reset: true,
+            ..Default::default()
+        }));
     }
 
-    let mut fg_colors = Vec::with_capacity(fg_args.len());
+    fg_colors.reserve(fg_args.len());
     for fg in fg_args {
         if is_builtin && fg == "reset" {
             continue;
         }
         if let Some(fg) = parse_color(fg)? {
+            if fg == Color::Normal {
+                reset = true;
+            }
             fg_colors.push(fg);
         }
     }
@@ -404,6 +430,7 @@ pub(crate) fn parse_text_face_and_options<'argarray, 'args>(
         bg,
         underline_color,
         style,
+        reset,
     }))
 }
 
@@ -424,7 +451,8 @@ mod tests {
                 })),
                 bg: None,
                 underline_color: None,
-                style: None
+                style: None,
+                reset: false
             }
         );
     }