+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Notable improvements and fixes
 ------------------------------
 - New Spanish translations (:issue:`12489`).
 - ``set_color`` is able to turn off italics, reverse mode, strikethrough and underline individually (e.g. ``--italics=off``).
+- ``set_color`` learned the foreground (``--foreground`` or ``-f``) and reset (``--reset``) modifiers.
 
 For distributors and developers
 -------------------------------